refactor websocket specific app methods
